+++ src/main/java/com/prezyk/md/Main.java
@@ -1,17 +1,19 @@
 package com.prezyk.md;
 
+import java.math.BigDecimal;
+
 public class Main {
     public static void main(String[] args) {
         SimulationInput simulationInput = SimulationInput.builder()
-                .time(10)
-                .boxSize(200)
-                .moleculeRadius(10)
+                .time(BigDecimal.valueOf(10))
+                .boxSize(BigDecimal.valueOf(200))
+                .moleculeRadius(BigDecimal.valueOf(10))
                 .moleculesQuantity(50)
-                .timeStep(0.01)
-                .mass(0.01)
-                .epsilon(50)
-                .wallStiffness(100)
-                                                         .build();
+                .timeStep(BigDecimal.valueOf(0.01))
+                .mass(BigDecimal.valueOf(0.01))
+                .epsilon(BigDecimal.valueOf(50))
+                .wallStiffness(BigDecimal.valueOf(100))
+                .build();
 
         MolecularDynamics molecularDynamics = new MolecularDynamics(simulationInput);
         Simulation simulation = molecularDynamics.calculateSimulation();
